Fixing some more prototypes. Pointed out by Markku Laukkanen.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@2597 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/ucd-snmp/memory.c
@@ -373,9 +373,9 @@ int getswap(int rettype)
 
 unsigned char *var_extensible_mem(struct variable *vp,
 				  oid *name,
-				  int *length,
+				  size_t *length,
 				  int exact,
-				  int *var_len,
+				  size_t *var_len,
 				  WriteMethod **write_method)
 {
